changes for backup wallet logout flow
