+++ src/screens/BackupWallet/BackupPhraseValidate/BackupPhraseValidate.js
@@ -17,7 +17,8 @@
    with this program; if not, write to the Free Software Foundation, Inc.,
    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-import React, { useState } from 'react';
+
+import React from 'react';
 import { Image } from 'react-native';
 import Clipboard from '@react-native-community/clipboard';
 import SafeAreaView from 'react-native-safe-area-view';
@@ -33,7 +34,6 @@ import Button from 'components/core/Button';
 import MnemonicPhrase from 'components/MnemonicPhrase';
 import Toast from 'components/Toast';
 import Text from 'components/core/Text';
-import CheckAuth from 'components/CheckAuth';
 
 // Actions
 import { backupWalletAction } from 'actions/walletActions';
@@ -44,46 +44,20 @@ import { appFont, spacing, fontSizes } from 'utils/variables';
 
 // Types
 import type { Dispatch } from 'reducers/rootReducer';
-import type { WalletObject } from 'models/Wallet';
 
 type Props = {
   navigation: NavigationScreenProp<*>,
-  resetIncorrectPassword: () => void,
   backupWallet: () => void,
 };
 
 // Assets
 const walletBackupImage = require('assets/images/logo-wallet-backup.png');
 
-const BackupPhraseValidate = ({ navigation, backupWallet, resetIncorrectPassword }: Props) => {
-  const wallet = navigation.getParam('wallet', null);
+const BackupPhraseValidate = ({ navigation, backupWallet }: Props) => {
+  const wallet = navigation.getParam('unlockedwallet', null);
 
-  const [pinIsValid, setPinIsValid] = useState(!!wallet);
-  const [unlockedwallet, setUnlockedWallet] = React.useState<?WalletObject>(wallet);
-
-  const mnemonicPhrase = unlockedwallet?.mnemonic;
-  const walletPrivateKey = unlockedwallet?.privateKey;
-
-  const handleScreenDismissal = () => {
-    resetIncorrectPassword();
-    navigation.goBack(null);
-  };
-
-  const onPinValid = (walletPin: ?string, { mnemonic, privateKey }) => {
-    setPinIsValid(true);
-    setUnlockedWallet({ mnemonic: mnemonic?.phrase, privateKey });
-  };
-
-  if (!pinIsValid || !mnemonicPhrase) {
-    return (
-      <CheckAuth
-        revealMnemonic
-        enforcePin
-        onPinValid={onPinValid}
-        headerProps={{ onClose: handleScreenDismissal }}
-      />
-    );
-  }
+  const mnemonicPhrase = wallet?.mnemonic;
+  const walletPrivateKey = wallet?.privateKey;
 
   const handleCopyToClipboard = (copiedText: ?string, isPrivateKey?: boolean) => {
     Clipboard.setString(copiedText);
